requirements.txt
fastmcp
pydantic
PyGithub
GitPython
pytest